chore: Split workflow_types.go into different files
